backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0

# Data validation
pydantic>=2.12.0,<3.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.9

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
hypothesis>=6.100.0

# Document processing
pymupdf>=1.24.0
python-docx>=1.1.0

# NLP (sadece sentence chunking için)
nltk>=3.8.0
langdetect>=1.0.9

# Evaluation metrics
rouge-score>=0.1.2

# API clients ve hesaplama
openai>=1.30.0
cohere>=5.0.0
numpy>=1.26.0
cachetools>=5.3.0
requests>=2.31.0

# Vector Database
weaviate-client>=4.9.0
